requirements.txt
openai
gradio
faiss-cpu
sentence-transformers
numpy
